# Performance fix in analyzer driver for CFG based analyzers